Unlock pages in CLEANUP rather than CLOSE
CLEANUP is called after the last handle is closed, whereas CLOSE is
called after the file itself is no longer in use. Since we want the
handle to be tied to the life time of its process, this ensures that the
kernel stops using the process memory before the process quits.

Or does it? Can't handles be duplicated across processes?

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>